[fe:fix] useGetMessage hook 두번호출 수정 - 댓글 컴포넌트 감싸는 걸로 수정 - Comments = Desktop + Mobile - useGetMessage => useEffect에서 처리
