+++ src/component/Section_Message/Modules/Comments/DesktopComments/index.module.scss
@@ -1,0 +1,15 @@
+@import '@style/_mixin.scss';
+@import '@style/_variables.scss';
+
+.commentLayout {
+  padding: 0px 20px;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  gap: 12px;
+  max-width: 813px;
+  height: 496px;
+  overflow-y: scroll;
+  -ms-overflow-style: none; /* 인터넷 익스플로러 */
+  scrollbar-width: none; /* 파이어폭스 */
+}
