+++ src/component/Section_Message/Modules/Comments/index.tsx
@@ -1,26 +1,32 @@
-import React from 'react';
-import { MessageType } from '@/types/dataType';
-import Comment from '@/component/Section_Message/Attom/Comment';
-import styles from './index.module.scss';
+import { useRecoilValue } from 'recoil';
 
-const Comments = ({ comments }: { comments: MessageType[] }) => {
+import useMessageData from '@/hooks/useGetMessage';
+import { deviceState } from '@/store';
+import styles from './index.module.scss';
+import MobileComments from './MobileComments';
+import DesktopComments from './DesktopComments';
+
+const Comments = () => {
+  const [comments, loading] = useMessageData();
+  const device = useRecoilValue(deviceState);
+  if (loading) {
+    return null;
+  }
+
+  console.log('hi');
   return (
-    <div className={styles.commentLayout}>
-      {comments.map(v => {
-        return (
-          <Comment
-            like_count={v.like_count}
-            key={v.message_id}
-            index={v.message_id}
-            nick={v.username}
-            date={v.created_dt}
-            isLike={v.user_liked}
-            content={v.content}
-            isDonate={false}
-          />
-        );
-      })}
-    </div>
+    <>
+      {device === 'desktop' && comments !== null && (
+        <div className={styles.commentsContainer}>
+          <DesktopComments comments={comments} />
+        </div>
+      )}
+      {device === 'mobile' && comments !== null && (
+        <div className={styles.commentsContainer}>
+          <MobileComments comments={comments} />
+        </div>
+      )}
+    </>
   );
 };
